# notebookOpenAPI.yaml
openapi: 3.0.3

info:
  title: Personal Notebook API
  description: REST API for managing personal notes.
  version: 1.0.0
  contact:
    email: user@server.com

tags:
  - name: Notes
    description: CRUD, search, and navigation operations on notes

paths:
  /notes:
    get:
      operationId: listNotes
      summary: List notes
      tags: [ Notes ]
      responses:
        '200':
          description: List of note metadata

    post:
      operationId: createNote
      summary: Create a new note
      tags: [ Notes ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteCreateRequest'
            examples:
              simple:
                summary: Simple text note
                value:
                  type: "simple"
                  title: "My First Note"
                  text: "Hello world!"
              list:
                summary: List note
                value:
                  type: "listnote"
                  title: "Shopping list"
                  list: [ "Milk", "Eggs", "Bread" ]
              bookmark:
                summary: Bookmark note
                value:
                  type: "bookmark"
                  title: "FastAPI Docs"
                  url: "https://fastapi.tiangolo.com"
      responses:
        '201':
          description: Note created successfully

  /notes/metadata:
    get:
      operationId: showMetadata
      summary: List notes (metadata only)
      tags: [ Notes ]
      responses:
        '200':
          description: List notes' metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  metadata:
                    description: The structure of the notes
                    type: object # (Change to 'array' if "get_note_structure returns()" a list)
                  message:
                    type: string
                    example: "Metadata showed"


  /notes/search:
    get:
      operationId: searchNotes
      summary: Search notes
      tags: [Notes]
      parameters:
        - name: title
          in: query
          description: Case-insensitive substring match on title
          schema:
            type: string
        - name: start_date
          in: query
          description: Start of creation date window
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: End of creation date window
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Matching notes

  /notes/{note_id}:
    parameters:
      - $ref: '#/components/parameters/NoteId'

    get:
      operationId: getNote
      summary: Get a specific note
      tags: [Notes]
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested note

    put:
      operationId: updateNote
      summary: Update a note
      tags: [ Notes ]
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpdateRequest'
            examples:
              simple:
                summary: Simple text note
                value:
                  type: "simple"
                  title: "My First Note"
                  text: "Hello world!"
              list:
                summary: List note
                value:
                  type: "listnote"
                  title: "Shopping list"
                  list: [ "Milk", "Eggs", "Bread" ]
              bookmark:
                summary: Bookmark note
                value:
                  type: "bookmark"
                  title: "FastAPI Docs"
                  url: "https://fastapi.tiangolo.com"
      responses:
        '200':
          description: The updated note

    delete:
      operationId: deleteNote
      summary: Delete a note by ID
      tags: [Notes]
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Note deleted

  /notes/{note_id}/navigate:
    parameters:
      - $ref: '#/components/parameters/NoteId'

    get:
      operationId: navigateNote
      summary: Fetch the content of a bookmark URL
      tags: [Notes]
      parameters:
        - name: note_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Raw content of the bookmarked URL

components:
  parameters:
    NoteId:
      name: note_id
      in: path
      required: true
      description: Unique numeric note identifier
      schema:
        type: integer
        format: int64

  schemas:
    NoteType:
      type: string
      enum: [ simple, listnote, bookmark ]
      description: The type of note to create or update

    NoteCreateRequest:
      type: object
      required: [ type, title ]
      properties:
        type:
          $ref: '#/components/schemas/NoteType'
        title:
          type: string
        text:
          type: string
        list:
          type: array
          items:
            type: string
        url:
          type: string
          format: uri
        tags:
          type: array
          items:
            type: string

    NoteUpdateRequest:
      type: object
      required: [ type ]
      properties:
        type:
          $ref: '#/components/schemas/NoteType'
        title:
          type: string
        text:
          type: string
        list:
          type: array
          items:
            type: string
        url:
          type: string
          format: uri